Did some refactoring to make the two folders interact with
