+++ Archive/Datasets/unify-preprocess-datasets.py
@@ -26,7 +26,7 @@ def convert_gtzan_to_h5(input_dir, output_dir):
 
 
 # Convert GTZAN to .h5
-convert_gtzan_to_h5('./genres', './gtzan_h5')
+convert_gtzan_to_h5('./genres', 'gtzan_h5')
 
 # Load the extracted features
 gtzan_df = pd.read_csv('gtzan_features.csv')
